refactor moving some functions from main.ts to each directory
