+++ src/github/index.ts
@@ -1,1 +1,2 @@
 export * from './github.js'
+export * from './types.js'
