+++ src/metrics/index.ts
@@ -1,1 +1,1 @@
-export * from './github.js'
+export * from './create-metrics.js'
